Move all utils to graph builder associated package
