+++ src/main/java/org/opentripplanner/graph_builder/module/transit/tickets/AvailableTicketsBuilderModule.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import org.opentripplanner.graph_builder.services.GraphBuilderModule;
 import org.opentripplanner.pricing.transit.ticket.TransitTicket;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.updater.transit.ticket.AvailableTransitTicketsGetter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
