--- src/test/java/org/opentripplanner/updater/transit/ticket/AvailableTransitTicketUpdaterTest.java
@@ -1,56 +1,0 @@
-package org.opentripplanner.updater.transit.ticket;
-
-import org.junit.Test;
-import org.opentripplanner.routing.graph.Graph;
-
-import static org.junit.Assert.assertEquals;
-
-public class AvailableTransitTicketUpdaterTest {
-
-    @Test
-    public void shouldAdd7TicketsToGraph() {
-        System.setProperty("ticketsDefinitionsFile", "src/test/resources/tickets/warsaw/availableTickets.json");
-        Graph graph = new Graph();
-
-        AvailableTransitTicketsUpdater updater = new AvailableTransitTicketsUpdater();
-        updater.setup(graph);
-        updater.configure(graph, null);
-        updater.runPolling();
-
-        assertEquals(7, graph.getAvailableTransitTickets().size());
-    }
-
-    @Test
-    public void shouldNotAddAnyTicketToGraph() {
-        System.clearProperty("ticketsDefinitionsFile");
-        Graph graph = new Graph();
-
-        AvailableTransitTicketsUpdater updater = new AvailableTransitTicketsUpdater();
-        updater.setup(graph);
-        updater.configure(graph, null);
-        updater.runPolling();
-
-        assertEquals(0, graph.getAvailableTransitTickets().size());
-    }
-
-    @Test
-    public void shouldReplace7TicketsWith0AfterUpdate() {
-        System.setProperty("ticketsDefinitionsFile", "src/test/resources/tickets/warsaw/availableTickets.json");
-        Graph graph = new Graph();
-
-        AvailableTransitTicketsUpdater updater = new AvailableTransitTicketsUpdater();
-        updater.setup(graph);
-        updater.configure(graph, null);
-        updater.runPolling();
-
-        assertEquals(7, graph.getAvailableTransitTickets().size());
-
-        System.clearProperty("ticketsDefinitionsFile");
-        updater = new AvailableTransitTicketsUpdater();
-        updater.setup(graph);
-        updater.configure(graph, null);
-        updater.runPolling();
-
-        assertEquals(0, graph.getAvailableTransitTickets().size());
-    }
-}
